refactor : getLectureByCourseId 를 lecture module에서 course module로 이동
