+++ src/modules/course/course.controller.ts
@@ -44,15 +44,15 @@ export class CourseController {
 		return this.courseService.getCategory2();
 	}
 
-	@Get('cat2/:id')
-	getCat2ByCat1Id(@Param('id') id:number){
-		return this.courseService.getCat2ByCat1(id);
-	}
-	
 	@Get(':id')
 	getCourseById(@Param('id') id: number) {
 		return this.courseService.getCourseById(id);
 	}
+
+	@Get('lecture/:id')
+    getLecturesByCourseId(@Param('id') id:number){
+        return this.courseService.getLecturesByCourseId(id);
+    }
 
 	@Post()
 	createCourse(@Body() createCourseDto: CreateCourseDto) {
