+++ src/modules/lecture/lecture.controller.ts
@@ -5,10 +5,5 @@ import { LectureService } from './lecture.service';
 @ApiTags('/lectures')
 @Controller('lectures')
 export class LectureController {
-    constructor(private lectureService: LectureService){}
-
-    @Get(':id')
-    getLecturesByCourseId(@Param('id') id:number){
-        return this.lectureService.getLecturesByCourseId(id);
-    }
+    
 }
